Do not use socket flags from address strings
Remove support for specifying socket flags in address strings. The
rationale is that the flags are useful for the programmer, but not
for end users, and address strings are designed to be exposed to the
users as well.

Instead, accept flags for socket options into the "flags" parameter
to the various functions, and define the values for them as enum
constants.

At the same time, make SOCK_CLOEXEC + SOCK_NONBLOCK the default, to
encourage good practices.