Simplify Unix socket creation
Factor out the netdialunix() (socket+connect) and netannounce()
(socket+bind) options by moving the shared code into unixsocket()
in the same fashion as for inetsocket(): passing a pointer to the
function that is to be called after socket() succeeds, which can
be either listen() or bind().

While at it, remove all the net{dial,announce}{unix,inet}() functions
and arrange to call {unix,inet}socket() directly.