Consistently use -1 for system call failures checks
Be strict about checking only for the valud -1 as return value for
system calls which use it for signalling an error. This does not change
the behavior, but avoids mixing styles ("fd < 0" vs. "fd == -1", for
example) and is more correct and can potentially avoid bugs in oddball
platforms which return something negative different from -1 on success;
not that this should be common, but it does no harm in using the proper
check.